refactor(openai): extract Responses-API transport and streaming collaborators
Decompose the 3330-line ModelHandlerOpenAIResponse monolith by extracting
three cohesive collaborators, following the existing support/utils/types
layout. Behavior-preserving — the orchestration in createResponseImpl is
unchanged, it now delegates to the collaborators.

- responseStreamEvents.ts: stateless event/item type guards and the
  response-id extractor, shared by the transport and the stream processor.
- ResponseStreamProcessor.ts: owns per-request streaming state (thinking/
  output streams, emitted web-search IDs) and the event state-machine,
  finalization, and web-search emission. Used by both the WebSocket and
  HTTP streaming paths via a small injected deps interface.
- OpenAIResponseWebSocketTransport.ts: owns the persistent WebSocket
  connection lifecycle (open, keepalive, reconnect, close) and request
  execution, feeding events into a per-request ResponseStreamProcessor.

The handler keeps the WebSocket-mode decision and wires the collaborators
through a lazy getter and a createStreamProcessor() factory. Handler drops
from 3330 to 2839 lines.
