+++ src/agent/modelHandlers/README.md
@@ -10,7 +10,7 @@ stays at the root.
 | ------------- | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------ |
 | _root_        | `ModelHandler` (abstract base), `toolConversion`, `contextManagementConstants`, `modelHandlerValidation` — shared across all providers                                                                                                                                               |
 | `anthropic/`  | `modelHandlerAnthropic` + Anthropic-only helpers (`anthropicContextManagement`, `anthropicDocumentHandling`, `anthropicTools`)                                                                                                                                                       |
-| `openai/`     | `modelHandlerOpenAI`, `modelHandlerOpenAIResponse`, OpenAI helpers (`openAIMessageUtils`, `openAIResponseErrors`, `BaseReasoningStreamAggregator`), and the OpenAI-**compatible** providers that extend `ModelHandlerOpenAI` (`modelHandlerDashScope/DeepSeek/GLM/Kimi/MiniMax/XAI`) |
+| `openai/`     | `modelHandlerOpenAI`, `modelHandlerOpenAIResponse`, Responses-API collaborators (`OpenAIResponseWebSocketTransport`, `ResponseStreamProcessor`, `responseStreamEvents`), OpenAI helpers (`openAIMessageUtils`, `openAIResponseErrors`, `BaseReasoningStreamAggregator`), and the OpenAI-**compatible** providers that extend `ModelHandlerOpenAI` (`modelHandlerDashScope/DeepSeek/GLM/Kimi/MiniMax/XAI`) |
 | `google/`     | `modelHandlerGoogleGenAI`                                                                                                                                                                                                                                                            |
 | `openrouter/` | `modelHandlerOpenRouterNative`                                                                                                                                                                                                                                                       |
 | `support/`    | Cross-provider runtime collaborators (stream handling, media, proxy, usage, SDK error adapters)                                                                                                                                                                                      |
